# docs/swagger.yaml
definitions:
...
  models.SbatchRequest:
    properties:
      account:
        type: string
      array:
        type: string
      begin:
        type: string
      chdir:
        type: string
      cluster_constraint:
        type: string
      clusters:
        items:
          type: string
        type: array
      comment:
        type: string
      constraint:
        type: string
      container:
        type: string
      container_id:
        type: string
      contiguous:
        type: boolean
      core_spec:
        type: integer
      cores_per_socket:
        type: integer
      cpu_freq:
        type: string
      cpus_per_gpu:
        type: integer
      cpus_per_task:
        type: integer
      deadline:
        type: string
      delay_boot:
        type: integer
      dependency:
        type: string
      distribution:
        type: string
      error:
        type: string
      exclude_nodes:
        items:
          type: string
        type: array
      exclusive:
        type: string
      export:
        type: string
      export_file:
        type: string
      extra_node_info:
        type: string
      get_user_env:
        type: boolean
      gid:
        type: string
      gpu_bind:
        type: string
      gpu_freq:
        type: string
      gpus:
        type: integer
      gpus_per_node:
        type: integer
      gpus_per_socket:
        type: integer
      gpus_per_task:
        type: integer
      gres:
        type: string
      gres_flags:
        type: string
      hint:
        type: string
      hold:
        type: boolean
      host:
        type: string
      ignore_pbs:
        type: boolean
      input:
        type: string
      job_name:
        type: string
      licenses:
        type: string
      mail_type:
        type: string
      mail_user:
        type: string
      mcs_label:
        type: string
      mem_bind:
        type: string
      mem_per_cpu:
        type: string
      mem_per_gpu:
        type: string
      memory:
        type: string
      min_cpus:
        type: integer
      nice:
        type: integer
      no_kill:
        type: boolean
      no_requeue:
        type: boolean
      node_file:
        type: string
      node_list:
        items:
          type: string
        type: array
      nodes:
        type: string
      ntasks:
        type: integer
      ntasks_per_core:
        type: integer
      ntasks_per_node:
        type: integer
      ntasks_per_socket:
        type: integer
      output:
        type: string
      overcommit:
        type: boolean
      oversubscribe:
        type: boolean
      parsable:
        type: boolean
      partition:
        type: string
      password:
        type: string
      port:
        type: integer
      power:
        type: string
      priority:
        type: integer
      privatekey:
        type: string
      profile:
        type: string
      propagate:
        type: string
      qos:
        type: string
      quiet:
        type: boolean
      reboot:
        type: boolean
      requeue:
        type: boolean
      reservation:
        type: string
      resv_ports:
        type: boolean
      script:
        type: string
      script_args:
        items:
          type: string
        type: array
      script_file:
        type: string
      signal:
        type: string
      sockets_per_node:
        type: integer
      spread_job:
        type: boolean
      switches:
        type: string
      thread_spec:
        type: integer
      threads_per_core:
        type: integer
      time:
        type: string
      time_min:
        type: string
      tmp_disk:
        type: string
      tres_bind:
        type: string
      tres_per_task:
        type: string
      uid:
        type: string
      use_min_nodes:
        type: boolean
      username:
        type: string
      verbose:
        type: boolean
      wait:
        type: boolean
      wckey:
        type: string
      wrap:
        type: string
    type: object
  models.SinfoRequest:
    properties:
      all:
        type: boolean
      clusters:
        items:
          type: string
        type: array
      dead:
        type: boolean
      exact:
        type: boolean
      federation:
        type: boolean
      format:
        type: string
      format_long:
        type: string
      future:
        type: boolean
      hide:
        type: boolean
      iterate:
        type: integer
      json:
        type: string
      list_reasons:
        type: boolean
      local:
        type: boolean
      long:
        type: boolean
      noconvert:
        type: boolean
      node_centric:
        type: boolean
      nodes:
        items:
          type: string
        type: array
      noheader:
        type: boolean
      partitions:
        items:
          type: string
        type: array
      port:
        type: integer
      reservation:
        type: boolean
      responding:
        type: boolean
      sort:
        items:
          type: string
        type: array
      states:
        items:
          type: string
        type: array
      summarize:
        type: boolean
      verbose:
        type: boolean
      yaml:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/slurm/sbatch/job/:
    post:
      consumes:
      - application/json
      description: 使用sbatch命令提交作业
      parameters:
      - description: sbatch请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SbatchRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 使用sbatch命令提交作业
      tags:
      - 作业管理
  /api/v1/slurm/sinfo/cluster/:
    post:
      consumes:
      - application/json
      description: 获取集群信息
      parameters:
      - description: sinfo请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SinfoRequest'
      - description: SSH会话密钥
        example: '"tsh_a2e932b625c0d598db3800aa91b92016"'
        in: header
        name: sessionKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功，返回作业会计信息列表
          schema:
            properties:
              home_path:
                type: string
              session_key:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误或SSH连接失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取集群、分区和节点信息
      tags:
      - 作业管理
...
swagger: "2.0"